Separate all modules from server
